add astrometry.net compatibility
`anet` fails for TESS fields near the poles for unclear reasons,
probably linked to the field distortion.

`astrometry.net` does not fail, and offers a variety of benefits:

* its index files are understandable, because they're well-documented.
* it works on the above fields with the "4100" index files.
* it is free software, with many users.
* it can output cute images of the constellations in your image.

etc.

So this commit offers the user the option to use `astrometry.net`
instead of `anet`. (Nothing dependent on `anet` is broken).

This can be done either in the sequence:
`fistar -> fistar_to_xy -> parallel_astrometrydotnet with
useimagenotfistar==False`, or just directly on fits images with
`parallel_astrometrydotnet with useimagenotfistar==True`.

The latter produces the cute images, but has less stable source
extraction.