add camera col to astromrefs+photrefs sql tables
This is a necessity for TESS reductions.

It is an easily-solved annoyance for HATPI reductions (by default, the
camera ID is always 0).

At various phases in autoimagesub, this leads to a "camera" option where
there previously wasn't one. For HATPI, the added few characters in
e.g., the astromref/photref file names is annoying.  However it is much
better than needing to deal with insane amounts of if/elsing in the
code.  It also may some day be useful, if HATPI cameras ever need
separate identifiers from ccds.

This should change nothing in HATPI reductions (other than adding a
"camera" string to the reference filenames).