add tweaks needed for photometry off base catalog
for lots of TESS science (and probably thus HATPI science) it is better
to do forced photometry based on catalog positions than photometry based
on source-extracted positions.

for TESS, this is because the pixels are 20", so in crowded regions,
you would have a bad time.

this commit updates various kwargs and adds logic so that this works,
at least for a single CCD of TESS ETE6.

these are needed because the different types of source extraction
produce (x,y) lists with different formats.