ete6: tuning to reduce CCDs serially
Initially, it seemed like a plausible idea that reducing multiple CCDs
in parallel would work, since it's an embarassingly parallel problem.

However, any given CCD easily takes up phtess1's 40 cores to begin with,
since everything is already parallelized across frames.

In addition, reducing multiple CCDs in parallel floods the postgres
database with queries, and it rejects some of them, leading some of the
jobs to fail.  (Because there is a max number of simultaneous
connections).

It is better, at least for now, and perhaps for the main job, unless I
ever migrate to AWS, to just run individual CCDs in serial.