tessutils: add parallel masking and trimming
`parallel_mask_saturated_stars`:

Mask the saturated stars using the calibrated images. Overwrites
image with fitsh-aware mask in the header. Note the mask is not actually
applied to the pixel values until we need to visualize; the mask is
metadata.

`mask_dquality_flag_frame`:

Mask the entire frame if dquality flag from SPOC is raised. Overwrites
image with fitsh-aware mask in the header. Note the mask is not
actually applied to the pixel values until we need to visualize; the
mask is metadata.

`parallel_trim_get_single_extension`:

makes `from_CAL_to_fitsh_compatible` ~10x faster