TESS_reduction: add record_reduction_parameters
To explore the parameter space of tunable parameters in the reduction,
we'll need a record of which parameters we've tuned.

Using the "projectid" as a unique identifier for each "different
reduction" seems like a good solution.

Each projectid gets a pickle file (and log file) that can then be used
to compare statistics when turning the various knobs, and/or flicking
the various switches.