tessutils: implemented measure_known_HJ_SNR
this function:
* parses basic exoplanet archive information for a known HJ
* plots the lightcurves at various detrending stages
* runs bls
* runs a trapezoid fit to get depth and duration estimates
* estimates the noise from the out of transit flux in windows of +/-4
  transit durations around the transits
* makes plots of the trapezoid fits
* computes transit SNR given all the above, and writes it to a csv file

note:

I considered doing full Mandel-Agol transit model fitting for this step
(and actually had a working implementation for this).  However it was
much too expensive and slow, particularly since this entire pipeline is
currently on python2.7, which complicated the multiprocessing with
`emcee`. Since we only need a maximum-likelihood estimate of the transit
depth to estimate the SNR, the approach I implemented seems fast and
good enough.