add RMS vs mag plot ability, instead of MAD vs mag
although in the case of gaussian noise the RMS and MAD are cleanly
related, in the case of 1/f noise this is not true.

so sometimes we might be interested in a different estimator.

one benefit of MAD is that it is more outlier resistant.

one benefit of RMS is that it is what we developed the TESS noise model
to predict. (so for the TESS reduction, it is the easier thing to use to
compare achieved precision vs. predicted precision)