TESS EPD: add ccd temperature term
this is a work-in-progress commit; `epd_lightcurve_imagesub` will now
raise a NotImplementedError, because including the CCD temperature
column in the lightcurves requires a "change of format". This is a bit
silly (https://github.com/waqasbhatti/pipe-trex/issues/15); I will
finish this part of the commit once our lightcurve format has been
settled.