upgraded to python 3.X
There are some new astropy features that require python 3; scipy's new
releases will now (as Jan 1 2019) support python 3 only; numpy is a
similar story. Regardless, we'll be using this code for a while, and
upgrading to python 3 is sensible.

This breaks one dependency on HATpipepy (which is in python 2). This
dependency was `HATpipepy.Common.BinPhot.read_fiphot`, which was only
needed to read binary fiphot output.  Running reductions with text
fiphot output works around this problem.

The majority of edits in this commit deal with character encoding for
binary vs text streams, which python 3 made more explicit than 2. It
works for a TESS reduction, which is my version of a unit-test right
now.

For HATPI reductions (@waqasbhatti, @samuelyeewl), there may be some
steps that I missed.

(Perhaps implementing a real unit-testing framework with continuous
integration is reaching the point of feasibility as well...)