add TFA (and optionally, BLS+LS+killharm sequence)
this commit shifts the dependence for TFA from the HATpipe TFA to the
public vartools TFA.

the TFA appears to run for a randomly-selected "TUNE" projid.

one known bug: it seems that a fraction of the lightcurves that are
produced come out with a single defective point at their beginning.