lcutils: fix cross-device link bug
For some reason, about 1 in 1e6 FITS overwrites produced the following
error:

OSError: [Errno 18] Invalid cross-device link:
'/nfs/phtess1/ar1/TESS/FFI/LC/FULL/s0002/ISP_1-2-1153/2328258821488678784_llc.fits'

This seems to happen in a non-reproduceable way. In other words: if I
rerun the same code that produced the error, it won't necessarily
produce it again. (Scary). Perhaps it has to do with memory allocation
during read/writes.

Apparently such errors can happen when "os.rename" is being used on
NFS-mounted system, instead of "shutil.move". (the latter is file-system
independent, and uses shutils.copy2().)

The `_overwrite_existing` constructor in astropy,
https://github.com/astropy/astropy/blob/master/astropy/io/fits/file.py,
seems to use `os.remove` rather than `os.rename`, so it's not so clear
why this error ever gets raised.

However, a reasonable fix should be to remove files that will be
ovewritten anyway. (Since this is supposedly being done under the hood
anyway). That is what this commit implements, but it's not clear that it
will work.