tessutils: add parallel_bkgd_subtract
this would belong better in a module dedicated to image processing.

see the docstrings, but the idea here is that before image subtraction,
it is probably better to subtract out "scattered light" from
Earth/Moonlight.

for TESS, the scattered light is big blobs, and so it can be taken out
with a median box filter.