fix lcutils.merge_tfa_worker bug
VARTOOLS TFA treats NaNs differently from python, and returns
lightcurves with NaNs filtered out.

In other words, it returns lightcurves of different lengths than it
receives.

To fix this, the previous implementation used `np.insert` statements to
insert NaNs at the appropriate indices.

However, this failed in particular cases, because `np.insert` didn't
like it when the indices to insert *to* didn't exist in the original
lightcurves.

This commit gives a solution that actually works: make an array full of
NaNs. Then fill it with the appropriate TFA magnitudes (matched from
RSTFC IDs) that exist.