Various fixes for database I/O in autoimagesub.py
Note: in arefshifted_frame_to_database, the on conflict clause
has been updated to use (arefshiftedframe, astromref) as the unique
index. This constraint needs to be added to the database schema.

Previously, only arefshiftedframe was used as the unique index, but this
constraint did not exist in the database schema either, causing any
attempt to overwrite rows to fail.

The fix to get_frame_info helps avoid a failure mode where a fiphot
contains only a single row. TODO: figure out why this happens, and why
such frames do not get moved into bad_frames.