bug-fix reference flux measurement procedure contd
This commit does two things:

1) adds `--spline` option call to the `fiphot` call that performs
photometry on subtracted images.  This is because the aperture in
difference image photometry needs to be convolved, and re-normalized
(the left term on the RHS of Pal's 2009 thesis, Eqn 83).

2) removes default `magfit` call from the same.  The default, as it was
written, was unsafe b/c the `.cmrawphot` file being passed
(`photrefrawphot` in L324 of imagesubphot.py) did not necessarily
contain catalog magnitude or color information.
The exact details of magnitude fitting (e.g., Zhang+2016, section 5.5)
also depend enough on the type & quality of data that the default should
not be to include it.

If `domagfit` is set to true, a nominal addition to fiphot of:

    "--magfit orders=4:2,niter=3,sigma=3 "
    "--col-magnitude 5 --col-color 6 "

is included, but these numbers are randomly chosen and the user should
optimize them. (Or see if they make any difference).

Relevant issues:

  https://github.com/waqasbhatti/pipe-trex2/issues/9

  https://github.com/waqasbhatti/pipe-trex/issues/18