fixed bugs in PCA detrending
I wasn't thinking correctly about the linear regression step needed to compute
the coefficients, once the PCA to get the eigenvectors was done.  I got
confused, and thought that I needed to be using the PCA eigenvalues as weights,
rather than just solving for an independent set of coefficients given the PCA
eigenvectors.  Oops!

A few points:

* To find the coefficients that best fit the linear combination of
however many principal components are being used, ordinary least squares
does fine.

* Fancier is Bayesian Ridge Regression, which sounds extremely similar,
if not identical to the PDC-MAP procedure (except it's in
machine-learning jargon). It doesn't seem to change the resulting light
curves by much for the few random examples I plotted.

* To determine the "optimal" number of components, it seems like the
Factor Analysis cross-validation score is a good estimator. We might do
this independently for each (sector, camera, ccd), since presumably the
optimal number of components varies depending on the systematics.