wcsqualityassurance: require CTYPE keys intact
In sector 9, some of the frames are wrecked by Earth-shine, but the
quality flags in the FFIs do not reflect this. (...)

The failure mode for the pipeline is then that it notices the WCS
derived from the SPOC headers is faulty.

This implements a failsafe that if the above is true, we move the frames
to the "badframe" graveyard.