wcs quality: add background-heavy failsafe
For some of the Earth-light scattered frames in sector 9, the WCS
quality assurance was failing.

This is because the centroids were poorly measured in regions of bad
background level (not because the WCS itself was wrong).

I could tell this because the errors between measured & projected
positions were largest at the edges of the background-heavy regions, and
not inside.

See:

https://github.com/lgbouma/cdips/commit/194fd98ceeabe3d65177d415594a7fed53d7303f#diff-db54fe2ec97d37f1c26ce8c8b2665cf7

To circumvent the issue, this commit weakens the WCS quality check
by a) throwing out background-heavy stars, and b) lowering the 90th
percentile and standard deviation criteria for these frames.