sector 9: add bad window
The DRN say that "The use of Camera 1 in attitude control was disabled
at the start of both orbits due to strong scattered light signals".

Empirically, the LCs at the start of orbit 26 (the 2nd orbit of the
sector) look fine (maybe already masked in the downlink window?). At the
start of 25, they show super-strong ramps. So, adding the 25 bad time
window.