lcutils: fix struct.error multiprocessing bug
This one took a while to figure out, since as usual multiprocessing bugs
are ridiculous levels of cryptic.

The issue was that each subprocess was re-instantiating a 4-Gb
dataframe.  (The gaia catalog of stars on silicon).

The solution I opted for was to put the dataframe into GLOBAL memory.

(This came up for these fields because centered on the galactic plane
--> more stars).