sector 8: update badtimes.
Throwing out 1.5 days of data, to save myself at least 1.5 days of
debugging needed to solve the issue of why TJD ~= 1530 to 1531.5 is so
wonky. (Answer is probably linked to rough scattered light + gaps
messing up the detrending). Tried first to reduce cam1ccd1 without doing
this, and it was jank.