revert 4067cf2
For photometric reference frame WCS quality checking, it introduced a
shift by SCIROW/SCICOL, leading to the quality check to fail.

I don't understand why yet.

Maybe the `anrd2xy` output xylists for the photometric reference do not
have the shift, while the TRIMMED CAL frames do, or something similar.

Regardless, this needs to be fixed if the shift from `anrd2xy` to
`wcs-rd2xy` is going to happen.