update TFA n_min_tfa_templates 120 -> 20
Cycle-4 initial reduction has showed that for a few camera/ccd
combinations, typically camera 1 (for a northern ecliptic hemisphere
pointing), basically all the stars are quite variable.  Note also that
selecting "quiet" stars for TFA is (in my opinion) somewhat silly, as is
the practice of using arbitrarily an high number of TFA template stars
without regularization.  This change in some sense does modify the "TFA"
regressed light curves in the final CDIPS HLSP light curves.  However,
~all users, myself included, seem to prefer the RAW and PCA light
curves.  So the "dilution of product" implied by this change is, in my
view, not consequential.