imageutils: explicit typing to uint8 for jpegs
jpeg's do not encode floating point numbers.  Previously, when writing
the fits images to jpeg, the float to int conversion was performed
automatically, while raising a warning that it was lossy. An imageio
update changed this behavior, to raise an error.  Therefore we need to
explicitly do the type casting ourselves.