Update badframes for spoc-4.0.14 DQUALITY flags
EXP-TESS-ARC-ICD-TM-0014-Rev-F, and Cycle4 more generally, includes new
data quality flags.  Bits 13 (4096), 14 (8192), and 16 (32768) do not
apply to FFIs.

However bit 15 is a "Bad Calibration Exclude", which can be applied to
the FFIs.

For FFIs with this flag, it turns out that the nominal plate solution
fails, and CRVAL1 and CRVAL2 are not set.

While this is noted in the data quality tests, such frames should
generally be "badframes" in the CDIPS nomenclature.  For instance, they
produce errors if passed to ``given_fits_list_get_gain_exptime_ra_dec``.

This commit adds bit 15 as an "exclude" flag.

See:
https://archive.stsci.edu/files/live/sites/mast/files/home/missions-and-data/active-missions/tess/_documents/EXP-TESS-ARC-ICD-TM-0014-Rev-F.pdf